fix: rename StoryImage to StoryContainer
